refactor menu
